Extracting commons from OpenWhisk and our function
