+++ .github/configs/wordlist.txt
@@ -778,3 +778,4 @@ async
 AsyncMode
 AsyncResponseURL
 AsyncWaitToCollectMin
+Knative's
